feat(plugins): dedicated plugin manager dialog (#9548)
- Add PluginManagerDialog as the primary plugin-management surface, lifting
  install/enable/disable/uninstall/update-check/drag-drop out of the Settings
  PluginsTab into src/components/Plugin/
- Extract usePluginManager hook; single refreshKey-keyed load effect, reopen
  resets transient state without cross-cancelling the initial load
- Nested confirm + URL-input dialogs render at zIndex=nested for LIFO escape
- Add app.pluginManager action; open from the app menu (macOS + Win/Linux) and
  the thinned Plugins settings tab (installed-count CTA)
- Lazy-mount in App.tsx with hasOpened guard + idle prefetch; reserve a layout
  slot for the browse/discovery surface (#9305)
- Update settings search sections and move the management test suite to the dialog
