+++ src/components/Plugin/usePluginManager.ts
@@ -1,0 +1,528 @@
+import { useEffect, useRef, useState } from "react";
+import { useDeferredLoading } from "@/hooks";
+import { UI_DOHERTY_THRESHOLD } from "@/lib/animationUtils";
+import { formatErrorMessage } from "@shared/utils/errorMessage";
+import { logError } from "@/utils/logger";
+import type { LoadedPluginInfo, PluginInstallError } from "@shared/types/plugin";
+
+/** Enabled plugins first, then alphabetically by display name. */
+function sortPlugins(list: readonly LoadedPluginInfo[]): LoadedPluginInfo[] {
+  return [...list].sort((a, b) => {
+    const byEnabled = Number(a.disabled ?? false) - Number(b.disabled ?? false);
+    if (byEnabled !== 0) return byEnabled;
+    const aName = a.manifest.displayName ?? a.manifest.name;
+    const bName = b.manifest.displayName ?? b.manifest.name;
+    return aName.localeCompare(bName);
+  });
+}
+
+const UP_TO_DATE_NOTIFICATION_DURATION_MS = 4000;
+
+/**
+ * Map a structured install failure to user-facing copy. The bounded-fetch codes
+ * (F24) get tailored guidance; everything else falls back to the error's own
+ * message so manifest/engine failures still read clearly.
+ */
+function installErrorMessage(error: PluginInstallError | undefined): string {
+  switch (error?.code) {
+    case "fetch_timeout":
+      return "The download timed out. Check your connection and try again.";
+    case "size_exceeded":
+      return "That plugin is larger than the 30 MB limit.";
+    case "content_type_rejected":
+      return "That URL didn't return a plugin archive (.dntr). Check the URL and try again.";
+    case "fetch_failed":
+      return "Couldn't download the plugin. Check the URL and try again.";
+    default:
+      return error?.message ?? "Installation failed. Check the file and try again.";
+  }
+}
+
+type PendingUpdate = {
+  plugin: LoadedPluginInfo;
+  result: Extract<
+    Awaited<ReturnType<typeof window.electron.plugin.checkForUpdate>>,
+    { status: "available" }
+  >;
+};
+
+/**
+ * All plugin-management state and lifecycle handlers for the plugin manager
+ * dialog (#9548): list, install (file / URL / drop), enable/disable, uninstall,
+ * and the manual check-for-update flow. Extracted from the former Settings
+ * `PluginsTab` so the dialog stays a thin presentational consumer.
+ *
+ * `isOpen` is the dialog's visibility. A single data-loading effect keyed on
+ * `refreshKey` owns the `plugin.list()` pull; reopening the dialog bumps that
+ * counter rather than splitting into a second list-fetching effect that would
+ * cross-cancel the initial load (#4958).
+ */
+export function usePluginManager(isOpen: boolean) {
+  const [plugins, setPlugins] = useState<LoadedPluginInfo[]>([]);
+  const [loading, setLoading] = useState(true);
+  const showInlineLoading = useDeferredLoading(loading, UI_DOHERTY_THRESHOLD);
+  const [error, setError] = useState<string | null>(null);
+  const [notice, setNotice] = useState<string | null>(null);
+  const [pending, setPending] = useState<Set<string>>(new Set());
+  const [refreshKey, setRefreshKey] = useState(0);
+  const [pendingUninstall, setPendingUninstall] = useState<LoadedPluginInfo | null>(null);
+  // Defaults off so stored secrets survive a reinstall — the user opts in to
+  // wiping them. Reset whenever a new uninstall is armed so a prior tick can't
+  // carry over to the next plugin.
+  const [deleteSettings, setDeleteSettings] = useState(false);
+  const [isUninstalling, setIsUninstalling] = useState(false);
+
+  const armUninstall = (plugin: LoadedPluginInfo) => {
+    setDeleteSettings(false);
+    setPendingUninstall(plugin);
+  };
+
+  const closeUninstall = () => {
+    setPendingUninstall(null);
+    setDeleteSettings(false);
+  };
+  const [showUrlDialog, setShowUrlDialog] = useState(false);
+  const [urlInput, setUrlInput] = useState("");
+  const [isInstalling, setIsInstalling] = useState(false);
+  // Update-check state. `checkingUpdate` drives the per-row spinner; the ref is
+  // the synchronous reentrancy guard (state batches, leaving a double-click
+  // window — #4703). `upToDateId` shows the transient "Already up to date" note;
+  // `pendingUpdate` holds the fetched manifest preview for the confirm dialog.
+  const [checkingUpdate, setCheckingUpdate] = useState<Set<string>>(new Set());
+  const checkingUpdateRef = useRef<Set<string>>(new Set());
+  const [upToDateId, setUpToDateId] = useState<string | null>(null);
+  const upToDateTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
+  const [pendingUpdate, setPendingUpdate] = useState<PendingUpdate | null>(null);
+  const [isReinstalling, setIsReinstalling] = useState(false);
+  const [pendingHttpUrl, setPendingHttpUrl] = useState<string | null>(null);
+  // Integer depth counter so the overlay survives child enter/leave churn
+  // without flicker (same pattern as the terminal drop zone). The boolean is
+  // derived from it for the overlay render.
+  const dragDepthRef = useRef(0);
+  const [isDragOverFiles, setIsDragOverFiles] = useState(false);
+  // Synchronous mutex for the drop install loop. The `isInstalling` state lags
+  // a render, so two drops dispatched before React flushes could both read a
+  // stale `false`; the ref flips immediately and serializes them.
+  const installingRef = useRef(false);
+
+  // Clear the "Already up to date" auto-dismiss timer on unmount.
+  useEffect(() => {
+    return () => {
+      if (upToDateTimerRef.current) clearTimeout(upToDateTimerRef.current);
+    };
+  }, []);
+
+  // Single data-loading effect, keyed on `refreshKey`. Mutations re-pull by
+  // bumping the counter rather than splitting into a second effect that shares
+  // the `cancelled` flag (that would cross-cancel the initial load — #4958).
+  useEffect(() => {
+    let cancelled = false;
+    window.electron.plugin
+      .list()
+      .then((list) => {
+        if (cancelled) return;
+        setPlugins(sortPlugins(list));
+        setError(null);
+      })
+      .catch((err) => {
+        if (cancelled) return;
+        setError(formatErrorMessage(err, "Failed to load plugins"));
+        logError("Failed to load plugins", err);
+      })
+      .finally(() => {
+        if (!cancelled) setLoading(false);
+      });
+    return () => {
+      cancelled = true;
+    };
+  }, [refreshKey]);
+
+  // Reopening the dialog clears transient surfaces (errors, notices, open
+  // confirms) and re-pulls the authoritative list. The very first mount already
+  // loads via the `refreshKey` effect, so skip the bump then to avoid a
+  // duplicate fetch. This is the single reset path — it only bumps the counter,
+  // never fetches directly, so it can't cross-cancel the load effect (#4958).
+  const initialMountRef = useRef(true);
+  useEffect(() => {
+    if (!isOpen) return;
+    if (initialMountRef.current) {
+      initialMountRef.current = false;
+      return;
+    }
+    setError(null);
+    setNotice(null);
+    setUpToDateId(null);
+    setPendingUninstall(null);
+    setDeleteSettings(false);
+    setPendingUpdate(null);
+    setPendingHttpUrl(null);
+    setShowUrlDialog(false);
+    setUrlInput("");
+    setRefreshKey((k) => k + 1);
+  }, [isOpen]);
+
+  // Re-pull when any view installs or uninstalls a plugin (#9285 cross-view
+  // propagation). The event carries no payload, so always re-fetch the full
+  // list rather than patching from a stale closure (#5087).
+  useEffect(() => {
+    return window.electron.plugin.onProvenanceChanged(() => {
+      setRefreshKey((k) => k + 1);
+      // A plugin may have been uninstalled in another window — close any open
+      // reinstall confirm so it can't fire `installFromUrl` on a stale record.
+      setPendingUpdate(null);
+    });
+  }, []);
+
+  const handleToggle = async (plugin: LoadedPluginInfo) => {
+    const id = plugin.manifest.name;
+    if (pending.has(id)) return;
+    const next = plugin.disabled === true; // currently disabled → enabling
+    const wasDisabled = plugin.disabled;
+    const wasPendingRestart = plugin.pendingRestart;
+
+    setPending((prev) => new Set(prev).add(id));
+    // Optimistic flip: a single toggle always flips both the desired state and
+    // the running/desired mismatch, so `pendingRestart` simply inverts.
+    setPlugins((prev) =>
+      prev.map((p) =>
+        p.manifest.name === id ? { ...p, disabled: !next, pendingRestart: !p.pendingRestart } : p
+      )
+    );
+    try {
+      setError(null);
+      await window.electron.plugin.setEnabled(id, next);
+    } catch (err) {
+      // Revert only this plugin's fields with a functional updater — a
+      // whole-list snapshot could resurrect a plugin another view uninstalled
+      // (via onProvenanceChanged) while this toggle was in flight.
+      setPlugins((prev) =>
+        prev.map((p) =>
+          p.manifest.name === id
+            ? { ...p, disabled: wasDisabled, pendingRestart: wasPendingRestart }
+            : p
+        )
+      );
+      setError(formatErrorMessage(err, "Failed to update plugin"));
+      logError("Failed to update plugin enabled state", err);
+    } finally {
+      setPending((prev) => {
+        const copy = new Set(prev);
+        copy.delete(id);
+        return copy;
+      });
+    }
+  };
+
+  const handleInstallResult = (
+    result: Awaited<ReturnType<typeof window.electron.plugin.installFromFile>>
+  ) => {
+    switch (result.status) {
+      case "installed":
+        setError(null);
+        setNotice(null);
+        setRefreshKey((k) => k + 1);
+        return;
+      case "cancelled":
+        return;
+      case "not-implemented":
+        setNotice("Installing plugins isn't available yet — it lands with an upcoming release.");
+        return;
+      case "invalid-url":
+        setError("That doesn't look like a valid URL. Check it and try again.");
+        return;
+      case "failed": {
+        setNotice(null);
+        setError(installErrorMessage(result.errors[0]));
+        return;
+      }
+    }
+  };
+
+  const handleInstallFromFile = async () => {
+    if (isInstalling) return;
+    setIsInstalling(true);
+    try {
+      const result = await window.electron.plugin.installFromFile();
+      handleInstallResult(result);
+    } catch (err) {
+      setError(formatErrorMessage(err, "Failed to install plugin"));
+      logError("Failed to install plugin from file", err);
+    } finally {
+      setIsInstalling(false);
+    }
+  };
+
+  const performInstallFromUrl = async (url: string) => {
+    setIsInstalling(true);
+    try {
+      const result = await window.electron.plugin.installFromUrl(url);
+      handleInstallResult(result);
+      // Keep the dialog open for URL-correctable failures so the user can edit
+      // in place: a malformed URL, a download error (404 / DNS), or a response
+      // that wasn't a plugin archive. Size/timeout failures and successful
+      // installs close it.
+      const correctable =
+        result.status === "invalid-url" ||
+        (result.status === "failed" &&
+          (result.errors[0]?.code === "fetch_failed" ||
+            result.errors[0]?.code === "content_type_rejected"));
+      if (!correctable) {
+        setShowUrlDialog(false);
+        setUrlInput("");
+      }
+    } catch (err) {
+      setError(formatErrorMessage(err, "Failed to install plugin"));
+      logError("Failed to install plugin from URL", err);
+    } finally {
+      setIsInstalling(false);
+    }
+  };
+
+  const handleInstallFromUrl = async () => {
+    const url = urlInput.trim();
+    if (!url || isInstalling) return;
+    // Tier D2: a plaintext HTTP download is unauthenticated and unencrypted, so
+    // gate it behind an explicit confirm before the request leaves the app. The
+    // IPC handler still accepts http:// — this is the renderer's risk surface.
+    let protocol: string;
+    try {
+      protocol = new URL(url).protocol;
+    } catch {
+      // Let the handler return invalid-url so the messaging stays consistent.
+      await performInstallFromUrl(url);
+      return;
+    }
+    if (protocol === "http:") {
+      // Stack avoidance: close the URL input and surface the confirm on its
+      // own. Cancelling reopens the input with the URL intact so the user can
+      // switch to https; confirming proceeds with the install.
+      setShowUrlDialog(false);
+      setPendingHttpUrl(url);
+      return;
+    }
+    await performInstallFromUrl(url);
+  };
+
+  const confirmHttpInstall = async () => {
+    if (!pendingHttpUrl) return;
+    const url = pendingHttpUrl;
+    setPendingHttpUrl(null);
+    await performInstallFromUrl(url);
+  };
+
+  const cancelHttpInstall = () => {
+    setPendingHttpUrl(null);
+    setShowUrlDialog(true);
+  };
+
+  // Install one or more dropped `.dntr` files sequentially through #9292's
+  // install lock — one dialog per file, no batch UX (#9295). Non-`.dntr` drops
+  // surface a quiet inline notice; an empty path (synthetic File object that
+  // `getDroppedFilePath` can't resolve) is a structured error, not a missing
+  // file.
+  const installDroppedFiles = async (files: File[]) => {
+    const dntrFiles = files.filter((f) => f.name.toLowerCase().endsWith(".dntr"));
+    if (dntrFiles.length === 0) {
+      setError(null);
+      setNotice("Only .dntr files can be installed.");
+      return;
+    }
+    if (installingRef.current) return;
+    installingRef.current = true;
+    setNotice(null);
+    setError(null);
+    setIsInstalling(true);
+    // Collect per-file failures so a later file's success can't silently clear
+    // an earlier file's error (the loop installs one file at a time).
+    const failures: string[] = [];
+    try {
+      for (const file of dntrFiles) {
+        const path = window.electron.plugin.getDroppedFilePath(file);
+        if (!path) {
+          failures.push(`${file.name} — couldn't read its location, try Install from file`);
+          continue;
+        }
+        try {
+          const result = await window.electron.plugin.installFromPath(path);
+          if (result.status === "failed") {
+            failures.push(`${file.name} — ${result.errors[0]?.message ?? "install failed"}`);
+          } else {
+            handleInstallResult(result);
+          }
+        } catch (err) {
+          failures.push(`${file.name} — ${formatErrorMessage(err, "install failed")}`);
+          logError("Failed to install plugin from drop", err);
+        }
+      }
+    } finally {
+      installingRef.current = false;
+      setIsInstalling(false);
+    }
+    if (failures.length > 0) setError(failures.join("; "));
+  };
+
+  const handleDragEnter = (e: React.DragEvent) => {
+    if (!e.dataTransfer.types.includes("Files")) return;
+    e.preventDefault();
+    e.stopPropagation();
+    dragDepthRef.current++;
+    if (dragDepthRef.current === 1) setIsDragOverFiles(true);
+  };
+
+  const handleDragOver = (e: React.DragEvent) => {
+    if (!e.dataTransfer.types.includes("Files")) return;
+    e.preventDefault();
+    e.stopPropagation();
+    e.dataTransfer.dropEffect = "copy";
+  };
+
+  const handleDragLeave = (e: React.DragEvent) => {
+    if (!e.dataTransfer.types.includes("Files")) return;
+    e.stopPropagation();
+    dragDepthRef.current--;
+    if (dragDepthRef.current <= 0) {
+      dragDepthRef.current = 0;
+      setIsDragOverFiles(false);
+    }
+  };
+
+  const handleDrop = (e: React.DragEvent) => {
+    if (!e.dataTransfer.types.includes("Files")) return;
+    e.preventDefault();
+    e.stopPropagation();
+    dragDepthRef.current = 0;
+    setIsDragOverFiles(false);
+    const files = Array.from(e.dataTransfer.files);
+    if (files.length > 0) void installDroppedFiles(files);
+  };
+
+  const confirmUninstall = async () => {
+    if (!pendingUninstall) return;
+    const id = pendingUninstall.manifest.name;
+    setIsUninstalling(true);
+    try {
+      setError(null);
+      await window.electron.plugin.uninstall(id, deleteSettings);
+      closeUninstall();
+      // Re-pull the authoritative list rather than splicing the closure (#5087).
+      setRefreshKey((k) => k + 1);
+    } catch (err) {
+      setError(formatErrorMessage(err, "Failed to uninstall plugin"));
+      logError("Failed to uninstall plugin", err);
+    } finally {
+      setIsUninstalling(false);
+    }
+  };
+
+  const handleCheckForUpdate = async (plugin: LoadedPluginInfo) => {
+    const id = plugin.manifest.name;
+    // Synchronous reentrancy guard — a rapid double-click would otherwise fire
+    // two downloads before the first render flips the spinner (#4703).
+    if (checkingUpdateRef.current.has(id)) return;
+    checkingUpdateRef.current.add(id);
+    setCheckingUpdate((prev) => new Set(prev).add(id));
+    // A fresh check supersedes any lingering "up to date" note for this plugin.
+    if (upToDateId === id) setUpToDateId(null);
+    try {
+      setError(null);
+      const result = await window.electron.plugin.checkForUpdate(id);
+      switch (result.status) {
+        case "up-to-date":
+          if (upToDateTimerRef.current) clearTimeout(upToDateTimerRef.current);
+          setUpToDateId(id);
+          upToDateTimerRef.current = setTimeout(
+            () => setUpToDateId(null),
+            UP_TO_DATE_NOTIFICATION_DURATION_MS
+          );
+          break;
+        case "available":
+          setPendingUpdate({ plugin, result });
+          break;
+        case "fetch-failed":
+          setError(`Couldn't check for an update: ${result.message}`);
+          break;
+        case "invalid-id":
+          // The button is gated on `originalUrl !== null`, so this is a bug, not
+          // a user-facing state — log it without a toast.
+          logError("Update check returned invalid-id", new Error(id));
+          break;
+      }
+    } catch (err) {
+      setError(formatErrorMessage(err, "Failed to check for update"));
+      logError("Failed to check plugin for update", err);
+    } finally {
+      checkingUpdateRef.current.delete(id);
+      setCheckingUpdate((prev) => {
+        const copy = new Set(prev);
+        copy.delete(id);
+        return copy;
+      });
+    }
+  };
+
+  const confirmReinstall = async () => {
+    if (!pendingUpdate) return;
+    const url = pendingUpdate.plugin.originalUrl;
+    if (!url) {
+      setPendingUpdate(null);
+      return;
+    }
+    setIsReinstalling(true);
+    try {
+      setError(null);
+      const result = await window.electron.plugin.installFromUrl(url);
+      setPendingUpdate(null);
+      handleInstallResult(result);
+    } catch (err) {
+      setError(formatErrorMessage(err, "Failed to reinstall plugin"));
+      logError("Failed to reinstall plugin from URL", err);
+    } finally {
+      setIsReinstalling(false);
+    }
+  };
+
+  const closeUrlDialog = () => {
+    setShowUrlDialog(false);
+    setUrlInput("");
+  };
+
+  return {
+    plugins,
+    loading,
+    showInlineLoading,
+    error,
+    notice,
+    pending,
+    pendingUninstall,
+    deleteSettings,
+    setDeleteSettings,
+    isUninstalling,
+    armUninstall,
+    closeUninstall,
+    confirmUninstall,
+    showUrlDialog,
+    setShowUrlDialog,
+    closeUrlDialog,
+    urlInput,
+    setUrlInput,
+    isInstalling,
+    handleInstallFromFile,
+    handleInstallFromUrl,
+    pendingHttpUrl,
+    confirmHttpInstall,
+    cancelHttpInstall,
+    checkingUpdate,
+    upToDateId,
+    handleCheckForUpdate,
+    pendingUpdate,
+    setPendingUpdate,
+    isReinstalling,
+    confirmReinstall,
+    handleToggle,
+    isDragOverFiles,
+    handleDragEnter,
+    handleDragOver,
+    handleDragLeave,
+    handleDrop,
+  };
+}
