+++ src/components/Settings/settingsTabRegistry.tsx
@@ -1356,7 +1356,7 @@ export const SETTINGS_REGISTRY = [
     importKind: "lazy",
     importer: importPluginsTab,
     LazyComponent: LazyPluginsTab,
-    searchNavDescription: "Install, enable, disable, and uninstall plugins",
+    searchNavDescription: "Open the plugin manager to install, enable, and update plugins",
     searchNavKeywords: [
       "integrations",
       "plugins",
@@ -1369,39 +1369,25 @@ export const SETTINGS_REGISTRY = [
     ],
     sections: [
       {
-        id: "plugins-install",
-        section: "Installed plugins",
-        title: "Install a plugin",
-        description: "Install a plugin from a file or a URL.",
+        id: "plugins-manage",
+        section: "Plugins",
+        title: "Plugin manager",
+        description:
+          "Open the plugin manager to install from a file or URL, enable or disable, uninstall, and check for updates.",
         keywords: [
           "install",
           "plugin",
           "plugins",
+          "manage",
           "from file",
           "from url",
           "sideload",
           "dntr",
-          "extension",
-          "addon",
-        ],
-      },
-      {
-        id: "plugins-installed",
-        section: "Installed plugins",
-        title: "Installed plugins",
-        description:
-          "Turn installed plugins on or off, or uninstall them. Changes take effect after restart.",
-        keywords: [
-          "plugin",
-          "plugins",
           "enable",
           "disable",
+          "uninstall",
           "extension",
           "addon",
-          "builtin",
-          "uninstall",
-          "provenance",
-          "turn off",
         ],
       },
     ],
