+++ src/services/actions/__tests__/actionDefinitions.adversarial.test.ts
@@ -274,6 +274,7 @@ function createCallbacks(overrides: Partial<ActionCallbacks> = {}): ActionCallba
   return {
     onOpenSettings: vi.fn(),
     onOpenSettingsTab: vi.fn(),
+    onOpenPluginManager: vi.fn(),
     onToggleSidebar: vi.fn(),
     onToggleFocusMode: vi.fn(),
     onFocusRegionNext: vi.fn(),
