+++ src/services/actions/__tests__/actionDefinitions.bootstrap.test.ts
@@ -4,6 +4,7 @@ function createCallbacks() {
   return {
     onOpenSettings: () => {},
     onOpenSettingsTab: () => {},
+    onOpenPluginManager: () => {},
     onToggleSidebar: () => {},
     onToggleFocusMode: () => {},
     onFocusRegionNext: () => {},
