+++ src/services/actions/__tests__/actionDefinitions.project-system-preferences.adversarial.test.ts
@@ -203,6 +203,7 @@ function createCallbacks(overrides: Partial<ActionCallbacks> = {}): ActionCallba
   return {
     onOpenSettings: vi.fn(),
     onOpenSettingsTab: vi.fn(),
+    onOpenPluginManager: vi.fn(),
     onToggleSidebar: vi.fn(),
     onToggleFocusMode: vi.fn(),
     onFocusRegionNext: vi.fn(),
