+++ src/services/actions/__tests__/actionDefinitions.test.ts
@@ -7,6 +7,7 @@ async function createRegistry() {
   return createActionDefinitions({
     onOpenSettings: () => {},
     onOpenSettingsTab: () => {},
+    onOpenPluginManager: () => {},
     onToggleSidebar: () => {},
     onToggleFocusMode: () => {},
     onFocusRegionNext: () => {},
