+++ src/services/actions/__tests__/portalCloseEscalation.test.ts
@@ -20,6 +20,7 @@ async function createRegistry() {
   return createActionDefinitions({
     onOpenSettings: () => {},
     onOpenSettingsTab: () => {},
+    onOpenPluginManager: () => {},
     onToggleSidebar: () => {},
     onToggleFocusMode: () => {},
     onFocusRegionNext: () => {},
