+++ src/services/actions/__tests__/terminalGetOutputAction.test.ts
@@ -50,6 +50,7 @@ async function createRegistry() {
   return createActionDefinitions({
     onOpenSettings: () => {},
     onOpenSettingsTab: () => {},
+    onOpenPluginManager: () => {},
     onToggleSidebar: () => {},
     onToggleFocusMode: () => {},
     onFocusRegionNext: () => {},
