+++ src/services/actions/definitions/__tests__/actionMetadata.test.ts
@@ -57,6 +57,7 @@ function createStubCallbacks(): ActionCallbacks {
   return {
     onOpenSettings: () => {},
     onOpenSettingsTab: () => {},
+    onOpenPluginManager: () => {},
     onToggleSidebar: () => {},
     onToggleFocusMode: () => {},
     onFocusRegionNext: () => {},
